update commands about oem start up script

+++ src/oem_poststart.sh
@@ -1,17 +1,19 @@
 #!/bin/sh
 
-# Wait before trying to activate the modem other it won't work
-# TODO: Find a more reliable way to confirm that the activation worked
+#
+# Delay before trying to enable the modem otherwise it can fail
+# Note: /dev/smd8 only works on LE910C1 chips, see the Telit LE91* manual for more details
+# TODO: Replace sleep with a more reliable method
+#
 sleep 30
 echo "Activating modem" > /dev/kmsg
-# Read from the modem interface to get feedback when connecting
-cat /dev/smd8 &
+cat /dev/smd8 > /dev/kmsg &
 MON_PID="$!"
-# Note /dev/smd8 only works on LE910C1 chips
-# See the Telit LE91* manual for more details
 (printf 'at#sgact=1,1\r' | socat - /dev/smd8) ||:
 sleep 2
-kill -9 "$MON_PID" ||:
+if [ -n "$MON_PID" ]; then
+    kill -9 "$MON_PID" ||:
+fi
 
 sleep 1
 /data/tedge/start.sh > /dev/kmsg
